Move dbmaintenance parsing functions to common module
DH-1425
