+++ datahub/dbmaintenance/management/commands/update_investment_project_estimated_land_date.py
@@ -1,8 +1,8 @@
 from logging import getLogger
 
 import reversion
-from rest_framework.fields import BooleanField, DateField, UUIDField
 
+from datahub.dbmaintenance.utils import parse_bool, parse_date, parse_uuid
 from datahub.investment.models import InvestmentProject
 from ..base import CSVBaseCommand
 
@@ -29,10 +29,10 @@ class Command(CSVBaseCommand):
 
     def _process_row(self, row, simulate=False, **options):
         """Process one single row."""
-        pk = _parse_uuid(row['id'])
+        pk = parse_uuid(row['id'])
         investment_project = InvestmentProject.objects.get(pk=pk)
-        allow_blank_estimated_land_date = _parse_bool(row['allow_blank_estimated_land_date'])
-        estimated_land_date = _parse_date(row['estimated_land_date'])
+        allow_blank_estimated_land_date = parse_bool(row['allow_blank_estimated_land_date'])
+        estimated_land_date = parse_date(row['estimated_land_date'])
 
         if (investment_project.allow_blank_estimated_land_date == allow_blank_estimated_land_date
                 and investment_project.estimated_land_date == estimated_land_date):
@@ -49,22 +49,3 @@ class Command(CSVBaseCommand):
                 update_fields=('estimated_land_date', 'allow_blank_estimated_land_date')
             )
             reversion.set_comment('Estimated land date migration correction.')
-
-
-def _parse_bool(value):
-    return _parse_value(value, BooleanField())
-
-
-def _parse_date(value):
-    return _parse_value(value, DateField())
-
-
-def _parse_uuid(value):
-    return _parse_value(value, UUIDField())
-
-
-def _parse_value(value, field):
-    if not value or value.lower().strip() == 'null':
-        return None
-
-    return field.to_internal_value(value)
